+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+### 1.0.0 2022-06-12 Added new functions
+Functions added
+* bet_calc() calculated bet payout
+* kelly_bet() calculates dollar amount to bet via Kelly Criterion
+* kelly() calculates % of bankroll to bet via Kelly Criterion
+* over_round() calculates over round of bet
+* parlay_calc() calculates payout of parlay bets
+* true_implied_prob() calculates true implied probability
+* true_probability() calculates true probability
+
+
 ### 0.0.1 2020-07-20 Initial release
 Functions added
 * convert_odds() converts odds to different types